+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+##[1.3.1] - 2019-08-30 WangXH <kurapica125@outlook.com>
+### Changed
+- ProcessInnerRequest will return the status code as the first value, then the redirected path or the json data.
+
 ##[1.2.15] - 2019-08-23 WangXH <kurapica125@outlook.com>
 ### Changed
 - Fix the error message stack level for un-supported type settings in `__Arguments__`.